In process of adding matrix multiplication for gate operations

+++ app/PauliX.cc
@@ -1,6 +1,7 @@
 #include "PauliX.h"
+#include "Qubit.h"
 
 PauliX::PauliX(int control, int target, int size):Gate(control, target, size){
-    std::vector<int> matrix = {0, 1, 1, 0};
+    std::vector<double> matrix = {0.0, 1.0, 1.0, 0.0};
     gate.setMatrix(matrix);
 }
